# Speed up Version#<=> ~20-50% when lengths differ