@simi 2025-03-15
> I added some logging to my app, and 38% of Gem::Version comparisons during bundle update have different segment lengths.

@skipkayhil just wondering, is there any real difference on any of common `gem`/`bundle` operations? Are you using `Gem::Version` extensively in your app?