@skipkayhil 2025-03-15
> > I added some logging to my app, and 38% of Gem::Version comparisons during bundle update have different segment lengths.
> 
> @skipkayhil just wondering, is there any real difference on any of common `gem`/`bundle` operations? Are you using `Gem::Version` extensively in your app?

Ah my fault, I mentioned it in the last PR but not this one. 

`bundle update` is taking ~ 1 minute for me (using `bundler 2.6.2`), and `Gem::Version#<=>` is ~27% of that time.

---

I started with #8559 and this PR because `Gem::Version#<=>` is the single most called method in my app's `bundle update`, but since then I've found some larger opportunities for improvement as well:

- jhawthorn/pub_grub#37
  - Reduces my app's `bundle update` by ~18%
- jhawthorn/pub_grub#38, 
  - Reduces my app's `bundle update` by ~30% (and `Gem::Version#<=>` drops to ~14% of total runtime, so still worth optimizing imo)