@simi 2025-03-15
@skipkayhil thanks for the info, definitelly worth to optimize.