@simi 2025-03-15
@deivid-rodriguez Would it be worth to put benchmark in separate file and add to the codebase? Could be start of benchmark folder. We can do in separate PR, no need to block this one.