@deivid-rodriguez 2025-03-17
> @deivid-rodriguez Would it be worth to put benchmark in separate file and add to the codebase? Could be start of benchmark folder. We can do in separate PR, no need to block this one.

Maybe? Unless we actually exercise it in CI and find a way to make it complain if a PR regresses performance, I'm not sure how useful it will be but I'm not necessarily opposed.