@skipkayhil 2025-03-17
> I have a super minor suggestion that I'm not even sure it improves the implementation. It avoid assigning lhs twice to the same value in the first iteration of the second step, but maybe incrementing the counter like that makes it a bit weird, so feel free to disregard:

I like this, the double `lhs` assignment was definitely something that crossed my mind.

> I also found another potential suggestion that should speed things up in most cases (except when comparing two equal versions but with different "shape", like 1.0.0 and 1.0). Maybe worth trying it in your app and see the effect but not a requirement to merge this of course:

Yeah I noticed this was pretty expensive too. I'll followup with a new PR after I get stats to see how frequently this happens since the performance drop for that case was pretty large 👍